refactor: 여행일기 카드 회전 애니메이션을 커스텀 훅으로 분리
